page_visits_counter/requirements.txt
Flask==1.1.2
google-cloud-firestore==2.3.1